Move ConcurrentRandom into its own file
